Move TypeScript types into their own file.
