+++ web-ui/jest.config.js
@@ -13,6 +13,7 @@ const customJestConfig = {
     "^@/components/(.*)$": "<rootDir>/components/$1",
     "^@/pages/(.*)$": "<rootDir>/pages/$1",
     "^@/lib/(.*)$": "<rootDir>/lib/$1",
+    "^@/types/(.*)$": "<rootDir>/types/$1",
   },
   testEnvironment: "jest-environment-jsdom",
 };
